chore: update codemirror 5 to latest and auto update enable
