+++ gulpfile.js/index.js
@@ -25,11 +25,11 @@ const _ = require('lodash');
 const fs = require('fs');
 const webserver = require('gulp-webserver');
 const { src, dest, series } = require('gulp');
-const mergeStream =   require('merge-stream');
+// removed require('merge-stream') node module. it gives wired glob behavior and some files goes missing
 const zip = require('gulp-zip');
-const rename = require("gulp-rename");
 const jsDocGenerate = require('./jsDocGenerate');
 const Translate = require("./translateStrings");
+const copyThirdPartyLibs = require("./thirdparty-lib-copy");
 
 function cleanDist() {
     return del(['dist']);
@@ -58,106 +58,6 @@ function cleanAll() {
 function makeDist() {
     return src('src/**/*')
         .pipe(dest('dist'));
-}
-
-/**
- * Add thirdparty libs copied to gitignore except the licence file.
- * @returns {Promise<PassThrough>}
- */
-async function copyThirdPartyLibs(){
-    return mergeStream(
-        // @phcode/fs
-        src(['node_modules/@phcode/fs/dist/virtualfs.js',
-            'node_modules/@phcode/fs/dist/virtualfs.js.map'])
-            .pipe(dest('src/phoenix')),
-        // jszip
-        src(['node_modules/jszip/dist/jszip.js'])
-            .pipe(dest('src/thirdparty')),
-        src(['node_modules/jszip/LICENSE.markdown'])
-            .pipe(rename("jsZip.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // underscore
-        src(['node_modules/underscore/underscore-min.js'])
-            .pipe(dest('src/thirdparty')),
-        src(['node_modules/underscore/LICENSE'])
-            .pipe(rename("underscore.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // bootstrap
-        src(['node_modules/bootstrap/dist/js/bootstrap.min.js', 'node_modules/bootstrap/dist/js/bootstrap.min.js.map',
-            'node_modules/bootstrap/dist/css/bootstrap.min.css', 'node_modules/bootstrap/dist/css/bootstrap.min.css.map'])
-            .pipe(dest('src/thirdparty/bootstrap')),
-        src(['node_modules/bootstrap/LICENSE'])
-            .pipe(rename("bootstrap.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // hilightjs
-        src(['node_modules/@highlightjs/cdn-assets/highlight.min.js'])
-            .pipe(dest('src/thirdparty/highlight.js')),
-        src(['node_modules/@highlightjs/cdn-assets/styles/*.*'])
-            .pipe(dest('src/thirdparty/highlight.js/styles')),
-        src(['node_modules/@highlightjs/cdn-assets/languages/*.*'])
-            .pipe(dest('src/thirdparty/highlight.js/languages')),
-        src(['node_modules/@highlightjs/cdn-assets/LICENSE'])
-            .pipe(rename("highlight.js.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // gfm-stylesheet
-        src(['node_modules/@pixelbrackets/gfm-stylesheet/dist/gfm.min.css'])
-            .pipe(dest('src/thirdparty/')), // AGPL 2.0 license addded to
-        // prettier
-        src(['node_modules/prettier/*.js'])
-            .pipe(dest('src/extensions/default/Phoenix-prettier/thirdParty')),
-        src(['node_modules/prettier/LICENSE'])
-            .pipe(rename("prettier.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // font-awesome
-        src(['node_modules/@fortawesome/fontawesome-free/css/*'])
-            .pipe(dest('src/thirdparty/fontawesome/css')),
-        src(['node_modules/@fortawesome/fontawesome-free/js/*'])
-            .pipe(dest('src/thirdparty/fontawesome/js')),
-        src(['node_modules/@fortawesome/fontawesome-free/webfonts/*'])
-            .pipe(dest('src/thirdparty/fontawesome/webfonts')),
-        src(['node_modules/@fortawesome/fontawesome-free/svgs/brands/*'])
-            .pipe(dest('src/thirdparty/fontawesome/svgs/brands')),
-        src(['node_modules/@fortawesome/fontawesome-free/svgs/regular/*'])
-            .pipe(dest('src/thirdparty/fontawesome/svgs/regular')),
-        src(['node_modules/@fortawesome/fontawesome-free/svgs/solid/*'])
-            .pipe(dest('src/thirdparty/fontawesome/svgs/solid')),
-        src(['node_modules/@fortawesome/fontawesome-free/LICENSE.txt'])
-            .pipe(rename("fontawesome.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // devicons https://devicon.dev/
-        src(['node_modules/devicon/devicon.min.css'])
-            .pipe(dest('src/thirdparty/devicon/')),
-        src(['node_modules/devicon/fonts/*.*'])
-            .pipe(dest('src/thirdparty/devicon/fonts/')),
-        src(['node_modules/devicon/LICENSE'])
-            .pipe(rename("devicon.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // mime-db
-        src(['node_modules/mime-db/db.json'])
-            .pipe(rename("mime-db.json"))
-            .pipe(dest('src/thirdparty')),
-        src(['node_modules/mime-db/LICENSE'])
-            .pipe(rename("mime-db.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // marked.js markdown rendering
-        src(['node_modules/marked/marked.min.js'])
-            .pipe(dest('src/extensions/default/Phoenix-live-preview/thirdparty')),
-        src(['node_modules/marked/LICENSE.md'])
-            .pipe(rename("marked.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // @floating-ui for notification ui widget  floating-ui.dom.umd.min.js
-        src(['node_modules/@floating-ui/core/dist/floating-ui.core.umd.min.js'])
-            .pipe(dest('src/thirdparty')),
-        src(['node_modules/@floating-ui/dom/dist/floating-ui.dom.umd.min.js'])
-            .pipe(dest('src/thirdparty')),
-        src(['node_modules/@floating-ui/core/LICENSE'])
-            .pipe(rename("floating-ui.markdown"))
-            .pipe(dest('src/thirdparty/licences/')),
-        // documentation
-        src(['node_modules/documentation/LICENSE'])
-            .pipe(rename("documentation.markdown"))
-            .pipe(dest('src/thirdparty/licences/'))
-);
 }
 
 function serve() {
@@ -195,25 +95,35 @@ function zipDefaultProjectFiles() {
         .pipe(dest('src/assets/default-project/'));
 }
 
-function zipSampleProjectFiles() {
-    return mergeStream(
-        src(['src/assets/sample-projects/bootstrap-blog/**'])
-            .pipe(zip('bootstrap-blog.zip'))
-            .pipe(dest('src/assets/sample-projects/')),
-        src(['src/assets/sample-projects/explore/**'])
-            .pipe(zip('explore.zip'))
-            .pipe(dest('src/assets/sample-projects/')),
-        src(['src/assets/sample-projects/HTML5/**'])
-            .pipe(zip('HTML5.zip'))
-            .pipe(dest('src/assets/sample-projects/')),
-        src(['src/assets/sample-projects/dashboard/**'])
-            .pipe(zip('dashboard.zip'))
-            .pipe(dest('src/assets/sample-projects/')),
-        src(['src/assets/sample-projects/home-pages/**'])
-            .pipe(zip('home-pages.zip'))
-            .pipe(dest('src/assets/sample-projects/'))
-    );
+// sample projects
+function zipSampleProjectBootstrapBlog() {
+    return src(['src/assets/sample-projects/bootstrap-blog/**'])
+        .pipe(zip('bootstrap-blog.zip'))
+        .pipe(dest('src/assets/sample-projects/'));
 }
+function zipSampleProjectExplore() {
+    return src(['src/assets/sample-projects/explore/**'])
+        .pipe(zip('explore.zip'))
+        .pipe(dest('src/assets/sample-projects/'));
+}
+function zipSampleProjectHTML5() {
+    return src(['src/assets/sample-projects/HTML5/**'])
+        .pipe(zip('HTML5.zip'))
+        .pipe(dest('src/assets/sample-projects/'));
+}
+function zipSampleProjectDashboard() {
+    return src(['src/assets/sample-projects/dashboard/**'])
+        .pipe(zip('dashboard.zip'))
+        .pipe(dest('src/assets/sample-projects/'));
+}
+function zipSampleProjectHomePages() {
+    return src(['src/assets/sample-projects/home-pages/**'])
+        .pipe(zip('home-pages.zip'))
+        .pipe(dest('src/assets/sample-projects/'));
+}
+let zipSampleProjectFiles = series(zipSampleProjectBootstrapBlog, zipSampleProjectExplore, zipSampleProjectHTML5,
+    zipSampleProjectDashboard, zipSampleProjectHomePages);
+
 
 function _updateConfigFile(config) {
     delete config.scripts;
@@ -264,20 +174,20 @@ function createJSDocs() {
 }
 
 function generateDocIndex() {
-    return new Promise(async (resolve)=>{
+    return new Promise(async (resolve)=>{ // eslint-disable-line
         await jsDocGenerate.generateDocIndex('docs/generatedApiDocs');
         resolve();
     });
 }
 
 function translateStrings() {
-    return new Promise(async (resolve)=>{
+    return new Promise(async (resolve)=>{ // eslint-disable-line
         await Translate.translate();
         resolve();
     });
 }
 
-exports.build = series(copyThirdPartyLibs, zipDefaultProjectFiles, zipSampleProjectFiles);
+exports.build = series(copyThirdPartyLibs.copyAll, zipDefaultProjectFiles, zipSampleProjectFiles);
 exports.clean = series(cleanDist);
 exports.reset = series(cleanAll);
 exports.releaseDev = series(cleanDist, exports.build, makeDist);
