+++ package.json
@@ -33,7 +33,6 @@
         "http-server": "14.1.0",
         "husky": "^7.0.4",
         "lodash": "^4.17.21",
-        "merge-stream": "^2.0.0",
         "readable-stream": "^3.6.0",
         "requirejs": "^2.3.6",
         "through2": "^4.0.2"
@@ -73,6 +72,7 @@
         "@pixelbrackets/gfm-stylesheet": "^1.1.0",
         "bootstrap": "^5.1.3",
         "browser-mime": "^1.0.1",
+        "codemirror": "^5.65.6",
         "devicon": "^2.15.1",
         "jszip": "^3.7.1",
         "marked": "^4.0.14",
